backend/requirements.txt
fastapi>=0.100.0
uvicorn>=0.15.0
python-dotenv>=0.19.0
google-genai
numpy>=1.21.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-multipart>=0.0.5
tiktoken>=0.5.1
